A few more changes to index.html and chatStyle.css
I moved the exit button to the menu div, so it does not disappear when
there is a lot of info and you need to scroll. I also changed the size
of body to be 35% of the html and adjusted the wrapper div accordingly.
I think this is a little cleaner